Update README.md
The S&W routing protocol is a two phase (spray and wait phase) flooding-based routing protocol proposed by Spyropoulos et al. The main aim of the S&W routing protocol is to control the unnecessary message replicas that happen in the case of flooding-base routing protocols such as Epidemic. This is achieved by setting the maximum number of messages a source node can spread in the network (forward to relay nodes) during the spray phase. In wait phase, all nodes having a copy of a message stores the message copy on their buffer and wait until they perform direct delivery of a message to the destination node when the opportunity avails itself.
For the design of the proposed energy-efficient S&W routing protocol, the same assumptions that were utilized in the original routing protocols are considered, but the following energy-efficient parameters are also introduced:

Minimum energy threshold (T): This is a value set by the network administrator (in Joules) to prevent relay nodes that have low energy level from being selected as next-hop nodes. Before a source node can forward a message copy to an encountered relay node, the energy level of the neighbor node is checked so that if it is less than the prescribed value T, the message copy will not be sent to that node. This mechanism assists in selecting only the nodes that have enough energy to deliver the message copy to the destination node or closer to the destination node. However, if the encountered relay node is the destination node, the message copy is forwarded directly to it regardless of the energy level of that node. 
One-hop acknowledgment mechanism: This is implemented to remove the copies of an already delivered message from the network and to prevent unnecessary transmission of such message. In this mechanism, when the message copy has been delivered to its destination, the destination node sends an acknowledgment message to the last encounter node that has sent the message. Both nodes update their acknowledgment tables and drop the message from their buffer to prevent it from further relaying. After a message has been dropped by the two nodes, both nodes update the acknowledgment table and flood their updated acknowledgment table to the network so that any node that comes in contact with a node having updated acknowledgment table can update their own acknowledgment table and delete the already delivered messages from their buffer.